fix: split NotificationService.cs to satisfy PNX004 one-type-per-file rule
Moves NotificationSeverityConverter and NotificationIdConverter to
c-sharp/JsonUtils/, and splits the four service types into individual
files in c-sharp/Services/. No behavior change; all 289 tests pass.
