+++ c-sharp/Services/NotificationSeverity.cs
@@ -1,0 +1,12 @@
+using System.Text.Json.Serialization;
+using Paranext.DataProvider.JsonUtils;
+
+namespace Paranext.DataProvider.Services;
+
+[JsonConverter(typeof(NotificationSeverityConverter))]
+public enum NotificationSeverity
+{
+    Info,
+    Warning,
+    Error,
+}
